Moving gym files into a dedicated gym folder Also added documentatioon for gym and made sure it appears on the website
