+++ habitat-baselines/habitat_baselines/rl/ver/environment_worker.py
@@ -24,9 +24,9 @@ import numpy as np
 
 from habitat import RLEnv, logger, make_dataset
 from habitat.config import read_write
-from habitat.core.gym_env_episode_count_wrapper import EnvCountEpisodeWrapper
-from habitat.core.gym_env_obs_dict_wrapper import EnvObsDictWrapper
-from habitat.utils.gym_definitions import make_gym_from_config
+from habitat.gym import make_gym_from_config
+from habitat.gym.gym_env_episode_count_wrapper import EnvCountEpisodeWrapper
+from habitat.gym.gym_env_obs_dict_wrapper import EnvObsDictWrapper
 from habitat_baselines.common.tensor_dict import NDArrayDict, TensorDict
 from habitat_baselines.rl.ver.queue import BatchedQueue
 from habitat_baselines.rl.ver.task_enums import (
@@ -148,7 +148,7 @@ class EnvironmentWorkerProcess(ProcessBase):
         )
         self.send_transfer_buffers.slice_keys(full_env_result.keys())[
             self.env_idx
-        ] = full_env_result
+        ] = full_env_result  # type:ignore[assignment]
         self.queues.inference.put(self.env_idx)
 
     def _step_env(self, action):
@@ -175,7 +175,9 @@ class EnvironmentWorkerProcess(ProcessBase):
         self._last_obs, reward, done, info = self._step_env(action)
 
         with self.timer.avg_time("enqueue env"):
-            self.send_transfer_buffers[self.env_idx] = dict(
+            self.send_transfer_buffers[
+                self.env_idx
+            ] = dict(  # type:ignore[assignment]
                 observations=self._last_obs,
                 rewards=reward,
                 masks=not done,
@@ -215,7 +217,7 @@ class EnvironmentWorkerProcess(ProcessBase):
         self._torch_transfer_buffers = torch_transfer_buffers
         self._torch_transfer_buffers["environment_ids"][
             self.env_idx
-        ] = self.env_idx
+        ] = self.env_idx  # type:ignore[assignment]
         acts = self._torch_transfer_buffers["actions"].numpy()
         assert isinstance(acts, np.ndarray)
         self.actions = acts
