+++ habitat-lab/habitat/__init__.py
@@ -4,7 +4,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from habitat.config import get_config
+from habitat.config import get_config, read_write
 from habitat.core.agent import Agent
 from habitat.core.benchmark import Benchmark
 from habitat.core.challenge import Challenge
@@ -17,24 +17,3 @@ from habitat.core.simulator import Sensor, SensorSuite, SensorTypes, Simulator
 from habitat.core.vector_env import ThreadedVectorEnv, VectorEnv
 from habitat.datasets import make_dataset
 from habitat.version import VERSION as __version__  # noqa: F401
-
-__all__ = [
-    "Agent",
-    "Benchmark",
-    "Challenge",
-    "Dataset",
-    "EmbodiedTask",
-    "Env",
-    "get_config",
-    "logger",
-    "make_dataset",
-    "Measure",
-    "Measurements",
-    "RLEnv",
-    "Sensor",
-    "SensorSuite",
-    "SensorTypes",
-    "Simulator",
-    "ThreadedVectorEnv",
-    "VectorEnv",
-]
