+++ habitat-lab/habitat/config/__init__.py
@@ -7,3 +7,5 @@
 
 from habitat.config.default import DictConfig, get_config
 from habitat.config.read_write import read_write
+
+__all__ = ["get_config", "read_write", "default_structured_configs"]
