+++ habitat-lab/habitat/gym/gym_definitions.py
@@ -11,14 +11,14 @@ from typing import TYPE_CHECKING, Any, List, Optional
 import gym
 from gym.envs.registration import register, registry
 
-import habitat
-import habitat.utils.env_utils
+from habitat import get_config, read_write
 from habitat.config.default import _HABITAT_CFG_DIR
 from habitat.config.default_structured_configs import (
     SimulatorSensorConfig,
     ThirdRGBSensorConfig,
 )
 from habitat.core.environments import get_env_class
+from habitat.utils.env_utils import make_env_fn
 
 if TYPE_CHECKING:
     from omegaconf import DictConfig
@@ -52,15 +52,13 @@ def make_gym_from_config(config: "DictConfig") -> gym.Env:
     assert (
         env_class is not None
     ), f"No environment class with name `{env_class_name}` was found, you need to specify a valid one with env_task"
-    return habitat.utils.env_utils.make_env_fn(
-        env_class=env_class, config=config
-    )
+    return make_env_fn(env_class=env_class, config=config)
 
 
-def add_sim_sensor_to_config(
+def _add_sim_sensor_to_config(
     config: "DictConfig", sensor: SimulatorSensorConfig
 ):
-    with habitat.config.read_write(config):
+    with read_write(config):
         sim_config = config.habitat.simulator
         default_agent_name = sim_config.agents_order[
             sim_config.default_agent_id
@@ -82,9 +80,9 @@ def _make_habitat_gym_env(
     if override_options is None:
         override_options = []
 
-    config = habitat.get_config(cfg_file_path, overrides=override_options)
+    config = get_config(cfg_file_path, overrides=override_options)
     if use_render_mode:
-        add_sim_sensor_to_config(config, ThirdRGBSensorConfig())
+        _add_sim_sensor_to_config(config, ThirdRGBSensorConfig())
     env = make_gym_from_config(config)
     return env
 
@@ -103,13 +101,13 @@ if "Habitat-v0" not in registry.env_specs:
     # Generic supporting general configs
     _try_register(
         id_name="Habitat-v0",
-        entry_point="habitat.utils.gym_definitions:_make_habitat_gym_env",
+        entry_point="habitat.gym.gym_definitions:_make_habitat_gym_env",
         kwargs={},
     )
 
     _try_register(
         id_name="HabitatRender-v0",
-        entry_point="habitat.utils.gym_definitions:_make_habitat_gym_env",
+        entry_point="habitat.gym.gym_definitions:_make_habitat_gym_env",
         kwargs={"use_render_mode": True},
     )
 
@@ -123,18 +121,18 @@ if "Habitat-v0" not in registry.env_specs:
         full_path = osp.join(gym_task_config_dir, fname)
         if not fname.endswith(".yaml"):
             continue
-        cfg_data = habitat.get_config(full_path)
+        cfg_data = get_config(full_path)
         gym_name = _get_gym_name(cfg_data)
         if gym_name is not None:
             # Register this environment name with this config
             _try_register(
                 id_name=gym_template_handle % gym_name,
-                entry_point="habitat.utils.gym_definitions:_make_habitat_gym_env",
+                entry_point="habitat.gym.gym_definitions:_make_habitat_gym_env",
                 kwargs={"cfg_file_path": full_path},
             )
 
             _try_register(
                 id_name=render_gym_template_handle % gym_name,
-                entry_point="habitat.utils.gym_definitions:_make_habitat_gym_env",
+                entry_point="habitat.gym.gym_definitions:_make_habitat_gym_env",
                 kwargs={"cfg_file_path": full_path, "use_render_mode": True},
             )
